requirements-dev.txt
-r requirements.txt
autopep8
pyhamcrest
pylint
pytest
requests_mock
webtest
